% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_theme.R
\name{apa_theme}
\alias{apa_theme}
\title{Format a table using APA style}
\usage{
apa_theme(ft)
}
\arguments{
\item{ft}{Table object.}
}
\description{
This function can be applied to both gtsummary and flextable objects, I think.
Maybe it will work on other things?
I should update this later.
}
