% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_cor.R
\name{mice_cor}
\alias{mice_cor}
\title{Make a correlation matrix from imputed data.}
\usage{
mice_cor(mira, x, title)
}
\arguments{
\item{mira}{A \code{mira} object.}

\item{x}{List of variables from your \code{mira} object. Must be \code{character} class.}

\item{title}{The title of your correlation matrix.}
}
\description{
This takes data from a \code{mira} object and turns the variables you want
into a correlation matrix. Make sure your variable names are pretty.
I'll try to add something like x/y axis varnames in the future.
}
