% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_df.R
\name{mice_df}
\alias{mice_df}
\title{Create a descriptive statistics table from imputed data.}
\usage{
mice_df(mids, v, title)
}
\arguments{
\item{mids}{A \code{mids} object.}

\item{v}{A variable from your \code{mids}.}

\item{title}{The title of your correlation matrix. Must be \code{character} class.}
}
\description{
This takes data from a \code{mids} object and turns the variables you specify
into table of descriptive statistics This is useful when handling imputed data.
}
\details{
The output will be a \code{flextable} formatted in APA style.

\strong{Don't use this yet!} Still working on it.
}
