% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_table.R
\name{pval_table}
\alias{pval_table}
\title{Create a "menu" of corrected p-values.}
\usage{
pval_table(..., digits, title)
}
\arguments{
\item{...}{Numeric, enter p-values here.}

\item{digits}{Numeric, amount of digits after decimal point when rounding values. Optional.}

\item{title}{Character vector, table title. Optional.}
}
\description{
This function turns a list of p-values into correct versions using \code{p.adjust}.
}
\details{
The output will be a \code{flextable} formatted in APA style.
}
