% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecloth.R
\docType{package}
\name{tablecloth}
\alias{tablecloth}
\title{tablecloth}
\description{
The \code{tablecloth} package contains functions to help create publication-ready tables,
even with non-standard data types (e.g., \code{mids}).
}
\details{
\if{html}{\figure{sticker.PNG}{options: width="20\%"}}
}
\author{
Julianna Calabrese
}
