% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_out.R
\name{clear_out}
\alias{clear_out}
\title{Apply clear background to ggplot.}
\usage{
clear_out(clr)
}
\arguments{
\item{clr}{Desired outline color.}
}
\description{
Applies a totally-clear/transparent background to a ggplot2 object as a theme.
}
