% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_cor.R
\name{mice_cor}
\alias{mice_cor}
\title{Create a correlation matrix from imputed data.}
\usage{
mice_cor(mids, vs, title)
}
\arguments{
\item{mids}{A \code{mids} object.}

\item{vs}{Variables from \code{mids}. Must be \code{character} class. E.g., \code{c("bmi", "chl")}.}

\item{title}{The title of your correlation matrix. Must be \code{character} class. Optional.}
}
\description{
This takes data from a \code{mids} object and turns the variables you specify
into a correlation matrix. This is useful when handling imputed data.
}
\details{
The output will be a \code{flextable} formatted in APA style.
}
