% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_df.R
\name{mice_df}
\alias{mice_df}
\title{Create a descriptive statistics table from imputed data.}
\usage{
mice_df(mids, vs, title, nm)
}
\arguments{
\item{mids}{A \code{mids} object.}

\item{vs}{Variables from \code{mids}. Must be \code{character} class. E.g., c("bmi", "chl").}

\item{title}{The title of your correlation matrix. Must be \code{character} class. Optional.}

\item{nm}{Preferred variable names. Must be \code{character} class. E.g., c("bmi", "chl"). Optional.}
}
\description{
This takes data from a \code{mids} object and turns the variables you specify
into table of descriptive statistics This is useful when handling imputed data.
}
\details{
The output will be a \code{flextable} formatted in APA style.
}
